Moved calendar compontents into own folder
