+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import './App.css';
-import { EmergencyResponsePortalCalendar } from './components/Calendar';
+import { EmergencyResponsePortalCalendar } from './components/calendar-components/Calendar';
 
 class App extends Component {
   render() {
