+++ src/components/calendar-components/SideDisplay.js
